pyproject.toml
[project]
name = "clinical-analytics"
version = "0.1.0"
description = "Multi-dataset clinical analytics platform"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pandas>=2.1.0",
    "polars>=0.19.0",
    "duckdb>=0.9.0",
    "ibis-framework[duckdb]>=10.0.0",
    "streamlit>=1.28.0",
    "statsmodels>=0.14.0",
    "pyyaml>=6.0",
    "lifelines>=0.27.0",
    "openpyxl>=3.1.0",  # Excel file support
    "pyreadstat>=1.2.0",  # SPSS file support
    "pymupdf>=1.23.0",  # PDF text extraction
    "sentence-transformers>=5.2.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "structlog>=24.0.0",
    # FastAPI backend dependencies (ADR010)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sse-starlette>=2.2.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "python-multipart>=0.0.12",  # For form data / file uploads
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
# Test files use camelCase naming per CLAUDE.md: test_unit_scenario_expectedBehavior
"tests/**/*.py" = ["N802", "N806"]
# Scripts and UI pages need sys.path manipulation before imports
"scripts/**/*.py" = ["E402"]
"src/clinical_analytics/ui/**/*.py" = ["E402"]
"test_*.py" = ["E402"]
"tests/eval/**/*.py" = ["E402"]  # Eval scripts need sys.path manipulation

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (require external services)",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.performance]
# Regression thresholds (percentage increase)
individual_test_threshold = 20.0
suite_threshold = 15.0
slow_test_threshold_seconds = 30.0
baseline_file = "tests/.performance_baseline.json"
data_file = "tests/.performance_data.json"

[tool.uv.workspace]
members = [
    "docs",
]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-mermaid2-plugin>=1.2.3",
    "mkdocstrings[python]>=1.0.0",
    "pymdown-extensions>=10.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
]
